+++ src/main/java/nextstep/member/AuthConfig.java
@@ -2,7 +2,7 @@ package nextstep.member;
 
 import nextstep.member.domain.LoginMember;
 import nextstep.member.ui.AuthenticationPrincipalArgumentResolver;
-import nextstep.security.service.JwtTokenProvider;
+import nextstep.security.application.JwtTokenProvider;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
