+++ src/main/java/nextstep/member/application/GithubOAuth2Service.java
@@ -7,7 +7,7 @@ import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import nextstep.member.payload.AccessTokenResponse;
 import nextstep.member.payload.GithubUserInfoResponse;
-import nextstep.security.service.JwtTokenProvider;
+import nextstep.security.application.JwtTokenProvider;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
