+++ src/main/java/nextstep/member/application/JwtTokenProviderImpl.java
@@ -2,7 +2,7 @@ package nextstep.member.application;
 
 import io.jsonwebtoken.*;
 import nextstep.member.domain.LoginMember;
-import nextstep.security.service.JwtTokenProvider;
+import nextstep.security.application.JwtTokenProvider;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
