+++ src/main/java/nextstep/member/ui/AuthenticationPrincipalArgumentResolver.java
@@ -2,7 +2,7 @@ package nextstep.member.ui;
 
 import nextstep.member.AuthenticationException;
 import nextstep.member.domain.LoginMember;
-import nextstep.security.service.JwtTokenProvider;
+import nextstep.security.application.JwtTokenProvider;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
